MeMiniuts/requirements.txt
Streamlit == 1.39.0
PyPDF2 == 3.0.1
python-docx == 1.2.0
